refactor: move Settings into settings.rs
