+++ components/restreamer/src/state.rs
@@ -1,9 +1,11 @@
 //! Application state.
 
 mod mixin;
+mod settings;
 
-pub use self::mixin::{
-    Delay, Mixin, MixinId, MixinSrcUrl, Volume, VolumeLevel,
+pub use self::{
+    mixin::{Delay, Mixin, MixinId, MixinSrcUrl, Volume, VolumeLevel},
+    settings::Settings,
 };
 
 use std::{
@@ -35,66 +37,6 @@ use uuid::Uuid;
 use crate::{display_panic, serde::is_false, spec, srs, Spec};
 use chrono::{DateTime, Utc};
 use std::collections::HashMap;
-
-/// Server's settings.
-///
-/// It keeps different settings not related to restreams but to whole server
-#[derive(Clone, Debug, Deserialize, Serialize, Eq, PartialEq)]
-pub struct Settings {
-    /// [`argon2`] hash of password which protects access to this application's
-    /// public APIs.
-    pub password_hash: Option<String>,
-
-    /// [`argon2`] hash of password which protects access to single output
-    /// application's public APIs.
-    pub password_output_hash: Option<String>,
-
-    /// Title for the server
-    /// It is used for differentiating servers on UI side if multiple servers
-    /// are used.
-    pub title: Option<String>,
-
-    /// Whether do we need to confirm deletion of inputs and outputs
-    /// If `true` we should confirm deletion, `false` - do not confirm
-    pub delete_confirmation: Option<bool>,
-
-    /// Whether do we need to confirm enabling/disabling of inputs or outputs
-    /// If `true` we should confirm, `false` - do not confirm
-    pub enable_confirmation: Option<bool>,
-}
-
-impl Settings {
-    /// Exports this [`Settings`] as a [`spec::v1::Settings`].
-    #[inline]
-    #[must_use]
-    pub fn export(&self) -> spec::v1::Settings {
-        spec::v1::Settings {
-            delete_confirmation: self.delete_confirmation,
-            enable_confirmation: self.enable_confirmation,
-            title: self.title.clone(),
-        }
-    }
-
-    // Applies the given [`spec::v1::Settings`] to this [`Settings`].
-    ///
-    pub fn apply(&mut self, new: spec::v1::Settings) {
-        self.title = new.title;
-        self.delete_confirmation = new.delete_confirmation;
-        self.enable_confirmation = new.enable_confirmation;
-    }
-}
-
-impl Default for Settings {
-    fn default() -> Settings {
-        Settings {
-            password_hash: None,
-            password_output_hash: None,
-            title: None,
-            delete_confirmation: Some(true),
-            enable_confirmation: Some(true),
-        }
-    }
-}
 
 /// Server's info
 #[derive(
